+++ NEWS.md
@@ -3,6 +3,7 @@
 ## Bug fixes
 * `plot_all()` uses correct level-2 %NA in title
 * `simWide`: case with no observed bmi values removed
+* `traceplot()`, `densplot()`: `ncol` and `nrow` now work with `use_ggplot = TRUE`
 
 ## Minor changes
 * `plot_all()` uses `xpd = TRUE` when printing text for character variables